Final Header Structure
Our Logo image is not working write now, we have to resize there pixel for expected design.
Other than the logo, the header structure is final from my side. If you have some other suggestion then contact me. and it just a structure so ignore its indentation of presentation on browser.

check it!!